added preSignedUrl and env variables
There is a new lambda function to serve presigned URLs, as the media (images) will be stored in a separate S3 bucket (there are now 3: web app, media and lambda functions).
Also, the environment variables need to be available for each lambda function. It looks like this: the code on the local machine sets environment variables. These get loaded in Terraform when doing terraform plan/apply. In return, the components (here lambda functions) need to set environment variables, which are being read from the Terraform env variables. Hope that makes sense.