some functions still not working
I can create an account and log in. I.e. Cognito is working. However, API calls are not. These adjustments are to mimic what's in the current AWS config from James (directly done in AWS).
Furthermore, despite acl = "private" for 2 of the 3 S3 buckets, they still have full public access. You need to be specific to disable that - default is allow all (which kind of runs against the usual approach of disallowing all unless specified otherwise)