cognito should work as the other implementation
It took me a while to find out how to use a username AND email for signup. The docs weren't very helpful mapping AWS to terraform. I made NO progress on getting the permissions sorted for the presignedurl function and any of the API lambdas. Getting a 404 when using presigned - means lack of Authorisation.